+++ app/app/handlers/forms/moderator/add_naming.py
@@ -5,19 +5,33 @@ from loader import bot, dp
 from states.base import BaseStates
 from states.tgbot_states import AddNaming
 from utils import const, get_data
+from dependency_injector.wiring import inject, Provide
+from app.services.application import ApplicationService
+from app.core.container import Container
+from app.models.application import Application
 
 
 @dp.callback_query_handler(text='skip',
                            state=[AddNaming.add_several_naming,
                                   AddNaming.edit])
-async def skip(query: types. CallbackQuery, state: FSMContext):
-    await state.update_data(field_six=None)
-    await bot.delete_message(query.message.chat.id, query.message.message_id)
-    await get_data.send_data(query=query, state=state)
-    new_kb = kb.sure().add(kb.exit_button)
-    await query.message.answer(const.SURE, reply_markup=new_kb)
-    await state.set_state(AddNaming.sure)
-    await query.answer()
+@inject
+async def skip(query: types. CallbackQuery, state: FSMContext,
+               application: ApplicationService = Provide[Container.application_service]):
+    data = await state.get_data()
+    if 'admin' not in data:
+        await state.update_data(field_six='---')
+        await bot.delete_message(query.message.chat.id, query.message.message_id)
+        await get_data.send_data(query=query, state=state)
+        new_kb = kb.sure().add(kb.exit_button)
+        await query.message.answer(const.SURE, reply_markup=new_kb)
+        await state.set_state(AddNaming.sure)
+        await query.answer()
+    else:
+        await application.update(data['admin'],
+                                 obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                         'field_six': '---'})
+        await query.message.answer(const.CHANGE_SUCCESS)
+        await state.finish()
 
 
 async def get_section_material(message: types.Message, state: FSMContext):
@@ -137,7 +151,9 @@ async def correct(query: types.CallbackQuery, state: FSMContext):
     await query.answer()
 
 
-async def edit(message: types.Message, state: FSMContext):
+@inject
+async def edit(message: types.Message, state: FSMContext,
+               application: ApplicationService = Provide[Container.application_service]):
     data = await state.get_data()
     point = data['change']
     if point == 'name':
@@ -154,22 +170,64 @@ async def edit(message: types.Message, state: FSMContext):
         await state.update_data(field_five=message.text)
     elif point == 'several_naming':
         await state.update_data(field_six=message.document.file_id)
-    await get_data.send_data(message=message, state=state)
-    new_kb = kb.sure().add(kb.exit_button)
-    await message.answer(const.SURE,
-                         reply_markup=new_kb)
-    await state.set_state(AddNaming.sure)
+    data = await state.get_data()
+    if 'admin' not in data:
+        await get_data.send_data(message=message, state=state)
+        new_kb = kb.sure().add(kb.exit_button)
+        await message.answer(const.SURE, reply_markup=new_kb)
+        await state.set_state(AddNaming.sure)
+    else:
+        if 'name' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'name': data['name']})
+        elif 'field_one' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_one': data['field_one']})
+        elif 'field_two' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_two': data['field_two']})
+        elif 'field_three' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_three': data['field_three']})
+        elif 'field_four' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_four': data['field_four']})
+        elif 'field_five' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_five': data['field_five']})
+        elif 'field_six' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_six': data['field_six']})
+        await message.answer(const.CHANGE_SUCCESS)
+        await state.finish()
 
 
 @dp.callback_query_handler(state=AddNaming.edit)
-async def get_role(query: types.CallbackQuery, state: FSMContext):
+@inject
+async def get_role(query: types.CallbackQuery, state: FSMContext,
+                   application: ApplicationService = Provide[Container.application_service]):
     await bot.delete_message(query.message.chat.id, query.message.message_id)
     await state.update_data(role=query.data)
-    await get_data.send_data(query=query, state=state)
-    new_kb = kb.sure().add(kb.exit_button)
-    await query.message.answer(const.SURE,
-                               reply_markup=new_kb)
-    await state.set_state(AddNaming.sure)
+    data = await state.get_data()
+    if 'admin' not in data:
+        await get_data.send_data(query=query, state=state)
+        new_kb = kb.sure().add(kb.exit_button)
+        await query.message.answer(const.SURE,
+                                   reply_markup=new_kb)
+        await state.set_state(AddNaming.sure)
+    else:
+        await application.update(data['admin'],
+                                 obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                         'role': data['role']})
+        await query.message.answer(const.CHANGE_SUCCESS)
+        await state.finish()
 
 
 def register(dp: Dispatcher):
